requirements.txt
pysimplegui==4.28.0
playsound==1.2.2
python-vlc==3.0.11115